Allow new dogs to be added through rest API

+++ src/main/java/uk/co/outrun/model/DogNum.java
@@ -12,9 +12,9 @@ public class DogNum {
 
     @Getter
     @Setter
-    public long seq;
+    public int seq;
 
-    public DogNum(long seq) {
+    public DogNum(int seq) {
         this.seq= seq;
     }
